add non-polyfilled entrypoint typings
