+++ lib/parser-non-polyfilled.d.ts
@@ -1,0 +1,2 @@
+import { id3Parser } from '../id3Parser';
+export = id3Parser;
